% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Combine all the per-subject data files in a directory.}
\usage{
compile(data_dir, regex_key, header_file)
}
\arguments{
\item{data_dir}{Name of directory}

\item{regex_key}{Optional regular expression to specify matches
(e.g., "MWP1*")}

\item{header_file}{Optional file containing column headers}
}
\description{
Combine all the per-subject data files in a directory.
}

