% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{\%nin\%}
\alias{\%nin\%}
\title{return x's that are not in y's
usage: filter(my_data, subj_id %nin% outliers)}
\usage{
x \%nin\% y
}
\description{
return x's that are not in y's
usage: filter(my_data, subj_id %nin% outliers)
}

